+++ iati/tests/test_resources.py
@@ -5,6 +5,7 @@ import pytest
 import iati.constants
 import iati.resources
 import iati.validator
+import iati.version
 
 
 class TestResources(object):
@@ -49,8 +50,9 @@ class TestResourceFolders(object):
         (iati.Version('1.01'), '1-01'),
         ('1', '1'),
         ('2', '2'),
-        (None, 'version-independent')
+        (iati.version.STANDARD_VERSION_ANY, 'version-independent')
     ])
+    @pytest.mark.latest_version('2.02')
     def test_folder_name_for_version(self, version, expected_version_foldername):
         """Check that expected components are present within folder paths."""
         path = iati.resources.folder_name_for_version(version)